pyproject.toml
[project]
name = "azure-accesstoken"
version = "0.1.0"
dependencies = [
    "requests",
    "PyJWT"
]
